requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
Django==5.0.8
djangorestframework==3.15.2
exceptiongroup==1.2.2
flake8==7.1.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.12
langchain-core==0.2.38
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.99
marshmallow==3.21.3
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.36.1
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.15.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
regex==2024.5.15
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
tenacity==8.5.0
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
